RDM-4014 Extracted mapping code from CasesService
