+++ src/shared/components/case-editor/services/cases.service.ts
@@ -3,12 +3,12 @@ import { Observable, throwError } from 'rxjs';
 import { AbstractAppConfig } from '../../../../app.config';
 import { plainToClass } from 'class-transformer';
 import { Headers } from '@angular/http';
-import { OrderService, HttpService, HttpErrorService } from '../../../services';
+import { HttpErrorService, HttpService, OrderService } from '../../../services';
 import { ShowCondition } from '../../../directives/conditional-show/domain/conditional-show.model';
 import { catchError, map, tap } from 'rxjs/operators';
-import { CaseView, CaseEventTrigger, CaseEventData, CasePrintDocument, Draft, CaseField } from '../../../domain';
-import { WizardPage, WizardPageField } from '../domain';
-import { ComplexFieldMask } from '../domain/wizard-page-field-complex-mask.model';
+import { CaseEventData, CaseEventTrigger, CasePrintDocument, CaseView, Draft } from '../../../domain';
+import { WizardPage } from '../domain';
+import { WizardPageFieldToCaseFieldMapper } from './wizard-page-field-to-case-field.mapper';
 
 @Injectable()
 export class CasesService {
@@ -34,7 +34,8 @@ export class CasesService {
     private http: HttpService,
     private appConfig: AbstractAppConfig,
     private orderService: OrderService,
-    private errorService: HttpErrorService
+    private errorService: HttpErrorService,
+    private wizardPageFieldToCaseFieldMapper: WizardPageFieldToCaseFieldMapper
   ) {}
 
   getCaseView(jurisdictionId: string,
@@ -221,75 +222,9 @@ export class CasesService {
 
     eventTrigger.wizard_pages.forEach((wizardPage: WizardPage) => {
       wizardPage.parsedShowCondition = new ShowCondition(wizardPage.show_condition);
-      wizardPage.case_fields = wizardPage.wizard_page_fields.map((wizardField: WizardPageField) => {
-        return this.wizardPageFieldToCaseFieldMapper(wizardField, eventTrigger.case_fields);
-      });
-      wizardPage.case_fields = this.orderService.deepSort(wizardPage.case_fields);
+      wizardPage.case_fields = this.orderService
+        .deepSort(this.wizardPageFieldToCaseFieldMapper.mapAll(wizardPage.wizard_page_fields, eventTrigger.case_fields));
     });
   }
 
-  private wizardPageFieldToCaseFieldMapper(wizardPageField: WizardPageField, caseFields: CaseField[]): CaseField {
-
-    let case_field: CaseField = caseFields.find(e => e.id === wizardPageField.case_field_id);
-    case_field.id = wizardPageField.case_field_id;
-    case_field.wizardProps = wizardPageField;
-    case_field.display_context = wizardPageField.display_context;
-    case_field.order = wizardPageField.order;
-
-    if (wizardPageField.display_context === 'COMPLEX' && wizardPageField.complex_field_mask_list) {
-
-      wizardPageField.complex_field_mask_list.forEach((complexFieldMask: ComplexFieldMask) => {
-        const caseFieldIds = complexFieldMask.complex_field_id.split('.');
-        let case_field_leaf: CaseField;
-
-        if (case_field.field_type.type === 'Collection' && case_field.field_type.collection_field_type.type === 'Complex') {
-          const [_, ...tail] = caseFieldIds;
-          case_field_leaf = this.getCaseFieldLeaf(tail, case_field.field_type.collection_field_type.complex_fields);
-        } else {
-          case_field_leaf = this.getCaseFieldLeaf(caseFieldIds, caseFields);
-        }
-
-        if (complexFieldMask.display_context !== 'HIDDEN') {
-          case_field_leaf.hidden = false;
-          case_field_leaf.display_context = complexFieldMask.display_context;
-          if (complexFieldMask.order) {
-            case_field_leaf.order = complexFieldMask.order;
-          }
-          if (complexFieldMask.label && complexFieldMask.label.length > 0) {
-            case_field_leaf.label = complexFieldMask.label;
-          }
-          if (complexFieldMask.hint_text && complexFieldMask.hint_text.length > 0) {
-            case_field_leaf.hint_text = complexFieldMask.hint_text;
-          }
-          if (complexFieldMask.show_condition && complexFieldMask.show_condition.length > 0) {
-            case_field_leaf.show_condition = complexFieldMask.show_condition;
-          }
-        } else {
-          case_field_leaf.hidden = true;
-        }
-      });
-    }
-
-    return case_field;
-  }
-
-  private getCaseFieldLeaf(caseFieldId: string[], caseFields: CaseField[]): CaseField {
-    const [head, ...tail] = caseFieldId;
-    if (caseFieldId.length === 1) {
-      let caseLeaf = caseFields.find(e => e.id === head);
-      if (!caseLeaf) {
-        throw new Error(`Cannot find leaf for caseFieldId ${caseFieldId.join('.')}`);
-      }
-      return caseLeaf;
-    } else if (caseFieldId.length > 1) {
-      let caseField = caseFields.find(e => e.id === head);
-      if (!caseField.field_type && !caseField.field_type.complex_fields) {
-        throw new Error(`field_type or complex_fields missing for ${caseFieldId.join('.')}`);
-      }
-      let newCaseFields = caseField.field_type.complex_fields;
-      return this.getCaseFieldLeaf(tail, newCaseFields)
-    } else {
-      throw new Error(`Cannot find leaf for caseFieldId ${caseFieldId.join('.')}`);
-    }
-  }
 }
